Split off all main control subpanel functions from dashboard.py and put them into main_control.py
