+++ subsystem/main_control/__init__.py
@@ -1,0 +1,3 @@
+from .main_control import MainControlPanel
+
+__all__ = ["MainControlPanel"]
